// src/generated/models/Cr9be_playersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr9be_playersaxm365_program = {
  100000000: 'School',
  100000001: 'Academy',
  100000002: 'IndividualTraining',
  100000003: 'Tournament',
  100000004: 'Camp'
} as const;
export type Cr9be_playersaxm365_program = keyof typeof Cr9be_playersaxm365_program;
export const Cr9be_playerscr9be_gender = {
  216260000: 'Male',
  216260001: 'Female'
} as const;
export type Cr9be_playerscr9be_gender = keyof typeof Cr9be_playerscr9be_gender;
export const Cr9be_playersstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr9be_playersstatecode = keyof typeof Cr9be_playersstatecode;
export const Cr9be_playersstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr9be_playersstatuscode = keyof typeof Cr9be_playersstatuscode;

export interface Cr9be_playersBase {
  axm365_program?: Cr9be_playersaxm365_program;
  "axm365_Team@odata.bind"?: string;
  "cr9be_Coach@odata.bind"?: string;
  cr9be_dateofbirth?: string;
  cr9be_enrolleddate?: string;
  cr9be_footballfederationenrolleddate?: string;
  cr9be_gender?: Cr9be_playerscr9be_gender;
  "cr9be_Generation@odata.bind"?: string;
  "cr9be_Member@odata.bind"?: string;
  cr9be_name: string;
  cr9be_number?: number;
  cr9be_picture?: string;
  cr9be_playerid: string;
  "cr9be_Position@odata.bind"?: string;
  cr9be_surname?: string;
  cr9be_terminationdate?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cr9be_playersstatecode;
  statuscode?: Cr9be_playersstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cr9be_players extends Cr9be_playersBase {
  axm365_programname?: string;
  axm365_teamname?: string;
  axm365_teamyominame?: string;
  cr9be_coachname?: string;
  cr9be_gendername?: string;
  cr9be_generationname?: string;
  cr9be_membername?: string;
  cr9be_memberyominame?: string;
  cr9be_picture_timestamp?: number;
  cr9be_picture_url?: string;
  cr9be_pictureid?: string;
  cr9be_positionname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  axm365_team?: object;
  _axm365_team_value?: string;
  cr9be_coach?: object;
  _cr9be_coach_value?: string;
  cr9be_generation?: object;
  _cr9be_generation_value?: string;
  cr9be_member?: object;
  _cr9be_member_value?: string;
  cr9be_position?: object;
  _cr9be_position_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}
export type Cr9be_playersImageColumnName = 'cr9be_picture';

export type Cr9be_playersUploadColumnName = Cr9be_playersImageColumnName;