// src/generated/models/Axm365_eventattendancesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Axm365_eventattendancesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Axm365_eventattendancesstatecode = keyof typeof Axm365_eventattendancesstatecode;
export const Axm365_eventattendancesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Axm365_eventattendancesstatuscode = keyof typeof Axm365_eventattendancesstatuscode;

export interface Axm365_eventattendancesBase {
  axm365_attended?: boolean;
  "axm365_Event@odata.bind"?: string;
  axm365_eventattendanceid: string;
  axm365_name: string;
  "axm365_Player@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Axm365_eventattendancesstatecode;
  statuscode?: Axm365_eventattendancesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Axm365_eventattendances extends Axm365_eventattendancesBase {
  axm365_attendedname?: string;
  axm365_eventname?: string;
  axm365_playername?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  axm365_event?: object;
  _axm365_event_value?: string;
  axm365_player?: object;
  _axm365_player_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}